requirements.txt
torch==2.4.1
torchvision==0.19.1
tqdm==4.66.5
transformers==4.44.2
numpy==2.1.1
scikit-learn==1.5.2
evo==1.30.0
hydra-core==1.3.2
torchtyping==0.1.5
lightning==2.4.0
